Add a11y test for side menu component
