+++ bedrock/base/templates/macros-protocol.html
@@ -328,7 +328,7 @@
 #}
 {% macro sidemenu_lists(arr, body_id='') -%}
   <nav class="mzp-c-sidemenu">
-    <section class="mzp-c-sidemenu-summary mzp-js-toggle" aria-controls="sidebar-menu">
+    <section class="mzp-c-sidemenu-summary mzp-js-toggle" aria-controls="sidebar-menu" data-testid="side-menu-toggle">
       <h3 class="mzp-c-sidemenu-label">{{ ftl('ui-menu') }}</h3>
       <ul>
         <li>{{ arr[0][0][2]|e }}</li>
@@ -341,7 +341,7 @@
         {% endfor %}
       </ul>
     </section>
-    <section class="mzp-c-sidemenu-main">
+    <section class="mzp-c-sidemenu-main" id="sidebar-menu" data-testid="side-menu-main">
       {% for menu in arr %}
 
         <h4 class="mzp-c-sidemenu-title {% if menu[0][1] == body_id and body_id != '' %}mzp-is-current{% endif %}">
